+++ demos/hashsum.c
@@ -15,6 +15,8 @@
 #define basename(x) x
 #endif
 
+#include "common.h"
+
 #if !defined(PATH_MAX) && defined(_MSC_VER)
 #include <windows.h>
 #define PATH_MAX MAX_PATH
@@ -26,8 +28,6 @@
          (x >= 'A' && x <= 'F') ? 'A' - 10 : \
             '\255')
 #define HEXOF(x) (x - _base(x))
-
-static void register_algs(void);
 
 static char* hashsum;
 
@@ -259,87 +259,6 @@ int main(int argc, char **argv)
       }
    }
    return EXIT_SUCCESS;
-}
-
-static void register_algs(void)
-{
-  int err;
-  LTC_UNUSED_PARAM(err);
-
-#ifdef LTC_TIGER
-  register_hash (&tiger_desc);
-#endif
-#ifdef LTC_MD2
-  register_hash (&md2_desc);
-#endif
-#ifdef LTC_MD4
-  register_hash (&md4_desc);
-#endif
-#ifdef LTC_MD5
-  register_hash (&md5_desc);
-#endif
-#ifdef LTC_SHA1
-  register_hash (&sha1_desc);
-#endif
-#ifdef LTC_SHA224
-  register_hash (&sha224_desc);
-#endif
-#ifdef LTC_SHA256
-  register_hash (&sha256_desc);
-#endif
-#ifdef LTC_SHA384
-  register_hash (&sha384_desc);
-#endif
-#ifdef LTC_SHA512
-  register_hash (&sha512_desc);
-#endif
-#ifdef LTC_SHA512_224
-  register_hash (&sha512_224_desc);
-#endif
-#ifdef LTC_SHA512_256
-  register_hash (&sha512_256_desc);
-#endif
-#ifdef LTC_SHA3
-  register_hash (&sha3_224_desc);
-  register_hash (&sha3_256_desc);
-  register_hash (&sha3_384_desc);
-  register_hash (&sha3_512_desc);
-#endif
-#ifdef LTC_RIPEMD128
-  register_hash (&rmd128_desc);
-#endif
-#ifdef LTC_RIPEMD160
-  register_hash (&rmd160_desc);
-#endif
-#ifdef LTC_RIPEMD256
-  register_hash (&rmd256_desc);
-#endif
-#ifdef LTC_RIPEMD320
-  register_hash (&rmd320_desc);
-#endif
-#ifdef LTC_WHIRLPOOL
-  register_hash (&whirlpool_desc);
-#endif
-#ifdef LTC_BLAKE2S
-  register_hash (&blake2s_128_desc);
-  register_hash (&blake2s_160_desc);
-  register_hash (&blake2s_224_desc);
-  register_hash (&blake2s_256_desc);
-#endif
-#ifdef LTC_BLAKE2B
-  register_hash (&blake2b_160_desc);
-  register_hash (&blake2b_256_desc);
-  register_hash (&blake2b_384_desc);
-  register_hash (&blake2b_512_desc);
-#endif
-#ifdef LTC_CHC_HASH
-  register_hash(&chc_desc);
-  if ((err = chc_register(register_cipher(&aes_enc_desc))) != CRYPT_OK) {
-     printf("chc_register error: %s\n", error_to_string(err));
-     exit(EXIT_FAILURE);
-  }
-#endif
-
 }
 
 /* $Source$ */
